# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * INTELTEC Production Control API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
  message?: string;
}

export interface SuccessResponse {
  success: boolean;
  message?: string;
}

export interface LoginBody {
  email: string;
  password: string;
}

export type UserRole = (typeof UserRole)[keyof typeof UserRole];

export const UserRole = {
  admin: "admin",
  apontador: "apontador",
  cliente: "cliente",
} as const;

export interface User {
  id: number;
  name: string;
  email: string;
  role: UserRole;
  clienteId?: number | null;
  ativo: boolean;
  createdAt: string;
}

export interface LoginResponse {
  token: string;
  user: User;
}

export type CreateUserBodyRole =
  (typeof CreateUserBodyRole)[keyof typeof CreateUserBodyRole];

export const CreateUserBodyRole = {
  admin: "admin",
  apontador: "apontador",
  cliente: "cliente",
} as const;

export interface CreateUserBody {
  name: string;
  email: string;
  password: string;
  role: CreateUserBodyRole;
  clienteId?: number | null;
  ativo?: boolean;
}

export type UpdateUserBodyRole =
  (typeof UpdateUserBodyRole)[keyof typeof UpdateUserBodyRole];

export const UpdateUserBodyRole = {
  admin: "admin",
  apontador: "apontador",
  cliente: "cliente",
} as const;

export interface UpdateUserBody {
  name?: string;
  email?: string;
  role?: UpdateUserBodyRole;
  clienteId?: number | null;
  ativo?: boolean;
  password?: string;
}

export interface Cliente {
  id: number;
  nomeRazaoSocial: string;
  nomeFantasia?: string | null;
  nomeInterno?: string | null;
  emailContato?: string | null;
  emailProdutoFinalizado?: string | null;
  ativo: boolean;
  createdAt: string;
}

export interface CreateClienteBody {
  nomeRazaoSocial: string;
  nomeFantasia?: string;
  nomeInterno?: string;
  emailContato?: string;
  emailProdutoFinalizado?: string;
  ativo?: boolean;
}

export interface UpdateClienteBody {
  nomeRazaoSocial?: string;
  nomeFantasia?: string;
  nomeInterno?: string;
  emailContato?: string;
  emailProdutoFinalizado?: string;
  ativo?: boolean;
}

export interface Produto {
  id: number;
  descricao: string;
  impresso: boolean;
  envelopado: boolean;
  ativo: boolean;
  createdAt: string;
}

export interface CreateProdutoBody {
  descricao: string;
  impresso?: boolean;
  envelopado?: boolean;
  ativo?: boolean;
}

export interface UpdateProdutoBody {
  descricao?: string;
  impresso?: boolean;
  envelopado?: boolean;
  ativo?: boolean;
}

export type ProducaoStatus =
  (typeof ProducaoStatus)[keyof typeof ProducaoStatus];

export const ProducaoStatus = {
  RECEBIDA: "RECEBIDA",
  EM_PROCESSAMENTO: "EM_PROCESSAMENTO",
  PROCESSADA: "PROCESSADA",
  EM_PRODUCAO: "EM_PRODUCAO",
  EMBALADA: "EMBALADA",
  FINALIZADA: "FINALIZADA",
  CANCELADA: "CANCELADA",
} as const;

export interface ProducaoWithCliente {
  id: number;
  clienteId: number;
  dataRecebimento: string;
  observacoes?: string | null;
  status: ProducaoStatus;
  cliente: Cliente;
  createdAt: string;
  updatedAt: string;
}

export interface ProducaoItemWithProduto {
  id: number;
  producaoId: number;
  produtoId: number;
  itemNumero: number;
  quantidade: number;
  multiplicador: number;
  impresso: boolean;
  envelopado: boolean;
  embalado: boolean;
  despachado: boolean;
  dataUltimoStatus?: string | null;
  produto: Produto;
  producao: ProducaoWithCliente;
  createdAt: string;
}

export interface ProducaoDetail {
  id: number;
  clienteId: number;
  dataRecebimento: string;
  observacoes?: string | null;
  status: ProducaoStatus;
  cliente: Cliente;
  items: ProducaoItemWithProduto[];
  createdAt: string;
  updatedAt: string;
}

export interface CreateProducaoBody {
  clienteId: number;
  dataRecebimento: string;
  observacoes?: string;
}

export interface UpdateProducaoBody {
  clienteId?: number;
  dataRecebimento?: string;
  observacoes?: string;
  status?: ProducaoStatus;
}

export interface CreateProducaoItemBody {
  produtoId: number;
  quantidade: number;
  multiplicador?: number;
}

export interface UpdateProducaoItemBody {
  produtoId?: number;
  quantidade?: number;
  multiplicador?: number;
}

export interface MarcarItemsBody {
  itemIds: number[];
}

export type DashboardResumoPorStatus = { [key: string]: number };

export interface DashboardResumo {
  totalOrdens: number;
  porStatus: DashboardResumoPorStatus;
  ordensHoje: number;
  ordensFinalizadasHoje: number;
}

export interface PendentesPorEtapa {
  impressao: number;
  envelopamento: number;
  embalagem: number;
  despacho: number;
  processamento: number;
}

export type ListProducaoParams = {
  status?: string;
  clienteId?: number;
  dataInicio?: string;
  dataFim?: string;
};