# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: INTELTEC Production Control API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginBody"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current authenticated user
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Current user info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /clientes:
    get:
      operationId: listClientes
      tags: [clientes]
      summary: List all clients
      security:
        - bearerAuth: []
      responses:
        "200":
          description: List of clients
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Cliente"
    post:
      operationId: createCliente
      tags: [clientes]
      summary: Create a new client (admin only)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateClienteBody"
      responses:
        "201":
          description: Client created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cliente"
  /clientes/{id}:
    get:
      operationId: getCliente
      tags: [clientes]
      summary: Get a client by ID
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Client found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cliente"
    put:
      operationId: updateCliente
      tags: [clientes]
      summary: Update a client (admin only)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateClienteBody"
      responses:
        "200":
          description: Client updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cliente"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error
...
    LoginBody:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
      required:
        - email
        - password
    LoginResponse:
      type: object
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/User"
      required:
        - token
        - user
    User:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [admin, apontador, cliente]
        clienteId:
          type: integer
          nullable: true
        ativo:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - email
        - role
        - ativo
        - createdAt
...
    Cliente:
      type: object
      properties:
        id:
          type: integer
        nomeRazaoSocial:
          type: string
        nomeFantasia:
          type: string
          nullable: true
        nomeInterno:
          type: string
          nullable: true
        emailContato:
          type: string
          nullable: true
        emailProdutoFinalizado:
          type: string
          nullable: true
        ativo:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - nomeRazaoSocial
        - ativo
        - createdAt
    CreateClienteBody:
      type: object
      properties:
        nomeRazaoSocial:
          type: string
        nomeFantasia:
          type: string
        nomeInterno:
          type: string
        emailContato:
          type: string
        emailProdutoFinalizado:
          type: string
        ativo:
          type: boolean
      required:
        - nomeRazaoSocial
    UpdateClienteBody:
      type: object
      properties:
        nomeRazaoSocial:
          type: string
        nomeFantasia:
          type: string
        nomeInterno:
          type: string
        emailContato:
          type: string
        emailProdutoFinalizado:
          type: string
        ativo:
          type: boolean
...